chore: DTO 네이밍 수정
