+++ src/main/java/gg/agit/konect/notice/service/NoticeService.java
@@ -9,9 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import gg.agit.konect.council.model.Council;
 import gg.agit.konect.council.repository.CouncilRepository;
 import gg.agit.konect.notice.dto.CouncilNoticesResponse;
-import gg.agit.konect.notice.dto.NoticeCreateRequest;
-import gg.agit.konect.notice.dto.NoticeResponse;
-import gg.agit.konect.notice.dto.NoticeUpdateRequest;
+import gg.agit.konect.notice.dto.CouncilNoticeCreateRequest;
+import gg.agit.konect.notice.dto.CouncilNoticeResponse;
+import gg.agit.konect.notice.dto.CouncilNoticeUpdateRequest;
 import gg.agit.konect.notice.model.CouncilNotice;
 import gg.agit.konect.notice.repository.CouncilNoticeRepository;
 
@@ -31,13 +31,13 @@ public class NoticeService {
         return CouncilNoticesResponse.from(councilNoticePage);
     }
 
-    public NoticeResponse getNotice(Integer id) {
+    public CouncilNoticeResponse getNotice(Integer id) {
         CouncilNotice councilNotice = councilNoticeRepository.getById(id);
-        return NoticeResponse.from(councilNotice);
+        return CouncilNoticeResponse.from(councilNotice);
     }
 
     @Transactional
-    public void createNotice(NoticeCreateRequest request) {
+    public void createNotice(CouncilNoticeCreateRequest request) {
         Council council = councilRepository.getById(1);
         CouncilNotice councilNotice = request.toEntity(council);
 
@@ -45,7 +45,7 @@ public class NoticeService {
     }
 
     @Transactional
-    public void updateNotice(Integer id, NoticeUpdateRequest request) {
+    public void updateNotice(Integer id, CouncilNoticeUpdateRequest request) {
         CouncilNotice councilNotice = councilNoticeRepository.getById(id);
         councilNotice.update(request.title(), request.content());
     }
